Don't render closed views (fixes handle warnings)
Apparently, those react-beautiful-dnd warnings were
caused by the fact that onClose() is called *after*
the view has already been removed from the DOM, so
dnd can't find the handles it's looking for (via
querySelectorAll()).  This change simply avoids
rerendering the view once it's closed, so that dnd
won't look for stuff it's not going to find.  (For
some reason, on the main branch, unmounting in
onClose() was apparently creating the same issue.)