Un-confuse Typescript about MouseEvents
Apparently, you can't assign a MouseEvent to a MouseEvent.  Who knew?
Seriously, though, some bits of the code are treating the click as a
plain DOM MouseEvent, and others as a React-specific MouseEvent.  So
no matter what I do, something has to have a cast.  This is the bit
most recently changed, so it gets the cast, at least until the menu
stuff can be refactored to component classes.