Make boardModifiers state-independent
React setState functions can take a function that will be
called with the old state.  This avoids needing the modifiers
to be regenerated for every change, and thus forcing all of
the board to be re-rendered.  (Sadly, this *still* isn't
enough to stop the whole board re-rendering, but without
the change, the whole board *will* re-render.)