Fix generating new component types every render
The factory functions were creating new component types each time
they were called, which made it impossible for React to skip
rendering (because it always saw a fresh virtual DOM with a new
component type in place of the old one).  As of this revision,
only the items in changed columns are re-rendered: lanes whose
items were unchanged do not re-render their items.

This makes it *almost* fast enough to keep up with my regular
typing speed in a second window on my large board.  It seems
like it *should* be possible to prevent the remaining items from
rendering, but further investigation is required to be clear.

Still, we are almost to the end of the tunnel for this branch!