Make Item a PureComponent so it doesn't re-render
Functional components *always* re-render when their parents
do.  PureComponent classes only re-render when their props
or state change.  This could be further optimized by tweaking
shouldComponentUpdate() so that changes to laneIndex or itemIndex
don't cause a re-render (which would speed up refeshes when items
or lanes are moved), but the base algorithm is enough to keep up
with typing (unless stopped by GC, which there's still a lot of).