Window-specific timers (Fix #607)
This patch changes timers and animation frames to be as window-specific
as practical, both by targeting where possible, and shimming where not
(since 3rd party libraries also use these functions).  It fixes the
issue with card drag-drop freezing or stuttering when the board is in a
secondary window but the main window isn't visible.  (When the main
window isn't visible, RAF is blocked entirely and setTimeouts are
batched and delayed, among other restrictions and throttling.)