Introduced feature for tags titles and file properties
Added settings for adding the tags in title and/or porperties. Modified the 'New note from card' function to not remove the tags from card title, to use the settings to either leave the tags in the note title or not and use the settings to either add the tags to note properties or not.