Update Obsidian Kanban Plugin.md
Suggest adding this small note. 

I'm familiar with Kanban, but new to Obsidian. I was excited to try the Kanban feature, installed the plugin, but then had no idea how to make the board until I watched this video! https://forum.obsidian.md/t/kanban-plugin-video-tutorial/17444

This simple addition should make things smoother if this is a person's first plugin (or first plugin for a long time, my version of Obsidian has been languishing for years and I'm just getting back to it!)